✨ Allow answering polls from the calendar view
Make the calendar view (month + week) support casting votes, not just
viewing results. Clicking an option cycles yes → if-need-be → no, the
same interaction as the mobile poll.

- Extract the participant picker + submit controls out of the mobile
  poll into shared VotingControlBar / VotingSubmitButton, reused by
  both the mobile poll and the calendar.
- Week view: option cards become vote controls when answering, colored
  by the current vote.
- Day detail: reuses the mobile poll options for full cycle-to-vote
  rows while answering; stays a read-only breakdown otherwise.
- Add useOptionVote to read/write a single option's vote through the
  shared voting form.
