+++ apps/web/src/components/poll/mobile-poll.tsx
@@ -1,32 +1,17 @@
-import { Badge } from "@rallly/ui/badge";
-import { Button } from "@rallly/ui/button";
 import { Card, CardContent, CardFooter } from "@rallly/ui/card";
-import { Icon } from "@rallly/ui/icon";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@rallly/ui/select";
-import { MoreVerticalIcon, PlusIcon, UsersIcon } from "lucide-react";
 import { AnimatePresence } from "motion/react";
 import * as m from "motion/react-m";
 import type * as React from "react";
 import smoothscroll from "smoothscroll-polyfill";
 
 import { TimesShownIn } from "@/components/clock";
-import { OptimizedAvatarImage } from "@/components/optimized-avatar-image";
-import { Participant, ParticipantName } from "@/components/participant";
-import { ParticipantDropdown } from "@/components/participant-dropdown";
+import {
+  VotingControlBar,
+  VotingSubmitButton,
+} from "@/components/poll/voting-controls";
 import { useVotingForm } from "@/components/poll/voting-form";
-import { YouAvatar } from "@/components/poll/you-avatar";
 import { useOptions, usePoll } from "@/components/poll-context";
-import { usePermissions } from "@/contexts/permissions";
-import { Trans, useTranslation } from "@/i18n/client";
 
-import { useVisibleParticipants } from "../participants-provider";
-import { useUser } from "../user-provider";
 import GroupedOptions from "./mobile-poll/grouped-options";
 
 if (typeof window !== "undefined") {
@@ -34,137 +19,20 @@ if (typeof window !== "undefined") {
 }
 
 const MobilePoll: React.FunctionComponent = () => {
-  const pollContext = usePoll();
-
-  const { poll } = pollContext;
+  const { poll } = usePoll();
 
   const { options } = useOptions();
 
-  const session = useUser();
-
   const votingForm = useVotingForm();
-  const { formState } = votingForm;
 
   const selectedParticipantId = votingForm.watch("participantId");
-
-  const visibleParticipants = useVisibleParticipants();
-  const selectedParticipant = selectedParticipantId
-    ? visibleParticipants.find(
-        (participant) => participant.id === selectedParticipantId,
-      )
-    : undefined;
-
-  const { canEditParticipant, canAddNewParticipant } = usePermissions();
-
-  const { t } = useTranslation();
 
   const isEditing = votingForm.watch("mode") !== "view";
 
   return (
     <Card>
       <div className="flex flex-col space-y-2 border-b p-2">
-        <div className="flex gap-x-2.5">
-          {selectedParticipantId || !isEditing ? (
-            <Select
-              defaultValue="all"
-              value={selectedParticipantId}
-              onValueChange={(participantId) => {
-                votingForm.setValue("participantId", participantId);
-              }}
-              disabled={isEditing}
-            >
-              <SelectTrigger className="flex-1">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="all">
-                  <div className="flex items-center gap-x-2.5">
-                    <div>
-                      <Icon>
-                        <UsersIcon />
-                      </Icon>
-                    </div>
-                    <span>
-                      {t("allParticipants", {
-                        defaultValue: "All Participants",
-                      })}
-                    </span>
-                  </div>
-                </SelectItem>
-                {visibleParticipants.map((participant) => (
-                  <SelectItem key={participant.id} value={participant.id}>
-                    <Participant>
-                      <OptimizedAvatarImage
-                        name={participant.name}
-                        src={participant.image ?? undefined}
-                        size="sm"
-                      />
-                      <ParticipantName>{participant.name}</ParticipantName>
-                      {session.ownsObject(participant) && (
-                        <Badge>
-                          <Trans i18nKey="you" />
-                        </Badge>
-                      )}
-                    </Participant>
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          ) : (
-            <div className="flex grow items-center px-1">
-              <Participant>
-                <YouAvatar />
-                <ParticipantName>{t("you")}</ParticipantName>
-              </Participant>
-            </div>
-          )}
-          {isEditing ? (
-            <Button
-              onClick={() => {
-                if (votingForm.watch("mode") === "new") {
-                  votingForm.cancel();
-                } else {
-                  votingForm.setValue("mode", "view");
-                }
-              }}
-            >
-              {t("cancel")}
-            </Button>
-          ) : selectedParticipant ? (
-            <ParticipantDropdown
-              align="end"
-              disabled={!canEditParticipant(selectedParticipant.id)}
-              participant={{
-                name: selectedParticipant.name,
-                userId: selectedParticipant.userId ?? undefined,
-                email: selectedParticipant.email ?? undefined,
-                id: selectedParticipant.id,
-              }}
-              onEdit={() => {
-                votingForm.setEditingParticipantId(selectedParticipant.id);
-              }}
-            >
-              <Button
-                aria-label={t("moreOptions", { defaultValue: "More options" })}
-                size="icon"
-              >
-                <MoreVerticalIcon />
-              </Button>
-            </ParticipantDropdown>
-          ) : canAddNewParticipant ? (
-            <Button
-              aria-label={t("addParticipant", {
-                defaultValue: "Add participant",
-              })}
-              size="icon"
-              onClick={() => {
-                votingForm.newParticipant();
-              }}
-            >
-              <PlusIcon />
-            </Button>
-          ) : null}
-        </div>
+        <VotingControlBar />
       </div>
       {poll.options[0]?.duration !== 0 && poll.timeZone ? (
         <CardContent className="border-b">
@@ -199,16 +67,7 @@ const MobilePoll: React.FunctionComponent = () => {
             }}
           >
             <CardFooter className="border-t">
-              <Button
-                form="voting-form"
-                className="w-full"
-                type="submit"
-                variant="primary"
-                size="lg"
-                loading={formState.isSubmitting}
-              >
-                {selectedParticipantId ? t("save") : t("continue")}
-              </Button>
+              <VotingSubmitButton className="w-full" />
             </CardFooter>
           </m.div>
         ) : null}
